HHH-18410 Hoist some state to AbstractAttributeMapping to avoid megamorphic call sites
