+++ hibernate-core/src/main/java/org/hibernate/metamodel/mapping/internal/AbstractAttributeMapping.java
@@ -7,9 +7,11 @@
 package org.hibernate.metamodel.mapping.internal;
 
 import org.hibernate.metamodel.mapping.AttributeMapping;
+import org.hibernate.metamodel.mapping.AttributeMetadata;
 import org.hibernate.metamodel.mapping.ForeignKeyDescriptor;
 import org.hibernate.metamodel.mapping.ManagedMappingType;
 import org.hibernate.metamodel.mapping.MappingType;
+import org.hibernate.property.access.spi.PropertyAccess;
 import org.hibernate.type.descriptor.java.JavaType;
 
 /**
@@ -18,20 +20,39 @@ import org.hibernate.type.descriptor.java.JavaType;
 public abstract class AbstractAttributeMapping implements AttributeMapping {
 	private final String name;
 	private final int fetchableIndex;
+	private final int stateArrayPosition;
 
 	private final ManagedMappingType declaringType;
+	private final AttributeMetadata attributeMetadata;
+	private final PropertyAccess propertyAccess;
 
-	public AbstractAttributeMapping(String name, int fetchableIndex, ManagedMappingType declaringType) {
+	public AbstractAttributeMapping(
+			String name,
+			int fetchableIndex,
+			ManagedMappingType declaringType,
+			AttributeMetadata attributeMetadata,
+			int stateArrayPosition,
+			PropertyAccess propertyAccess) {
 		this.name = name;
 		this.fetchableIndex = fetchableIndex;
 		this.declaringType = declaringType;
+		this.attributeMetadata = attributeMetadata;
+		this.stateArrayPosition = stateArrayPosition;
+		this.propertyAccess = propertyAccess;
 	}
 
 	/**
 	 * For Hibernate Reactive
  	 */
 	protected AbstractAttributeMapping(AbstractAttributeMapping original) {
-		this( original.name, original.fetchableIndex, original.declaringType );
+		this(
+				original.name,
+				original.fetchableIndex,
+				original.declaringType,
+				original.attributeMetadata,
+				original.stateArrayPosition,
+				original.propertyAccess
+		);
 	}
 
 	@Override
@@ -42,6 +63,21 @@ public abstract class AbstractAttributeMapping implements AttributeMapping {
 	@Override
 	public String getAttributeName() {
 		return name;
+	}
+
+	@Override
+	public AttributeMetadata getAttributeMetadata() {
+		return attributeMetadata;
+	}
+
+	@Override
+	public int getStateArrayPosition() {
+		return stateArrayPosition;
+	}
+
+	@Override
+	public PropertyAccess getPropertyAccess() {
+		return propertyAccess;
 	}
 
 	@Override
