+++ hibernate-core/src/main/java/org/hibernate/metamodel/mapping/internal/AbstractSingularAttributeMapping.java
@@ -26,8 +26,6 @@ public abstract class AbstractSingularAttributeMapping
 		extends AbstractStateArrayContributorMapping
 		implements SingularAttributeMapping {
 
-	private final PropertyAccess propertyAccess;
-
 	public AbstractSingularAttributeMapping(
 			String name,
 			int stateArrayPosition,
@@ -36,8 +34,7 @@ public abstract class AbstractSingularAttributeMapping
 			FetchOptions mappedFetchOptions,
 			ManagedMappingType declaringType,
 			PropertyAccess propertyAccess) {
-		super( name, attributeMetadata, mappedFetchOptions, stateArrayPosition, fetchableIndex, declaringType );
-		this.propertyAccess = propertyAccess;
+		super( name, attributeMetadata, mappedFetchOptions, stateArrayPosition, fetchableIndex, declaringType, propertyAccess );
 	}
 
 	public AbstractSingularAttributeMapping(
@@ -49,8 +46,7 @@ public abstract class AbstractSingularAttributeMapping
 			FetchStyle fetchStyle,
 			ManagedMappingType declaringType,
 			PropertyAccess propertyAccess) {
-		super( name, attributeMetadata, fetchTiming, fetchStyle, stateArrayPosition, fetchableIndex, declaringType );
-		this.propertyAccess = propertyAccess;
+		super( name, attributeMetadata, fetchTiming, fetchStyle, stateArrayPosition, fetchableIndex, declaringType, propertyAccess );
 	}
 
 	/**
@@ -58,12 +54,6 @@ public abstract class AbstractSingularAttributeMapping
 	 */
 	protected AbstractSingularAttributeMapping( AbstractSingularAttributeMapping original ) {
 		super( original );
-		this.propertyAccess = original.propertyAccess;
-	}
-
-	@Override
-	public PropertyAccess getPropertyAccess() {
-		return propertyAccess;
 	}
 
 	@Override
